requirements.txt
boto3>=1.34.0
elasticsearch==9.3.0
easyocr==1.7.2
fastapi==0.116.1
opencv-python-headless==4.10.0.84
pika==1.3.2
pydantic==2.11.7
PyMuPDF==1.27.2
requests==2.32.5
python-docx==1.2.0
sentence-transformers==5.1.0
uvicorn==0.35.0
uvicorn[standard]==0.35.0
google-genai==1.69.0
